# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M3th",
        "description": "math testing",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://node43.webte.fei.stuba.sk",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/tests/{test}/questions": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "List all questions for a test, with optional search and tags filter",
                "operationId": "578e0eaf781bf4dd7be17389c04aa469",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in assignment_en or assignment_sk",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "description": "Filter by tag IDs (array)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions for the test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Create and attach a new question to the test",
                "operationId": "eb3f669fec3a3a1cae4fc6d1c8581f01",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignment_sk",
                                    "assignment_en",
                                    "isMultiChoice",
                                    "answers",
                                    "tag_ids"
                                ],
                                "properties": {
                                    "assignment_sk": {
                                        "type": "string",
                                        "example": "Koľko je 2+2?"
                                    },
                                    "assignment_en": {
                                        "type": "string",
                                        "example": "What is 2+2?"
                                    },
                                    "isMultiChoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "answer_sk": {
                                                    "type": "string",
                                                    "example": "Štyri"
                                                },
                                                "answer_en": {
                                                    "type": "string",
                                                    "example": "Four"
                                                },
                                                "isCorrect": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created and attached",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tests/{test}/questions/{question}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get one question (with answers and tags) for this test",
                "operationId": "7f2e07ba8076529d083bcc31c1cdc4a0",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The question resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update a question attached to this test",
                "operationId": "8571039851527e997e7356c61f94422f",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignment_sk",
                                    "assignment_en",
                                    "isMultiChoice",
                                    "answers",
                                    "tag_ids"
                                ],
                                "properties": {
                                    "assignment_sk": {
                                        "type": "string",
                                        "example": "Koľko je 2+2?"
                                    },
                                    "assignment_en": {
                                        "type": "string",
                                        "example": "What is 2+2?"
                                    },
                                    "isMultiChoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "answer_sk": {
                                                    "type": "string",
                                                    "example": "Štyri"
                                                },
                                                "answer_en": {
                                                    "type": "string",
                                                    "example": "Four"
                                                },
                                                "isCorrect": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Detach and delete a question from a test",
                "operationId": "bb15745dc16416a3574fbd29eab28685",
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question detached and deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Question 1 removed and deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Question": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "assignment_sk": {
                        "type": "string"
                    },
                    "assignment_en": {
                        "type": "string"
                    },
                    "isMultiChoice": {
                        "type": "boolean"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "Answer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "answer_sk": {
                        "type": "string"
                    },
                    "answer_en": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name_sk": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}